fixed monetization fractional

+++ blps/laba3/src/main/java/org/example/dto/response/MonetizationMethodResponse.java
@@ -8,7 +8,6 @@ import org.example.entity.AdType;
 import org.example.entity.MethodStatus;
 import org.example.entity.MonetizationType;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -21,7 +20,7 @@ public class MonetizationMethodResponse {
     private MonetizationType type;
     private AdType adType;
     private String adName;
-    private BigDecimal subscriptionPrice;
+    private Integer subscriptionPrice;
     private MethodStatus status;
     private String tags;
     private LocalDateTime createdAt;
